monitoring/trace/requirements.txt
flask
opentelemetry-distro
